QCEF via KDE and Quadrature
This initial commit implements a scalar observation-space filter similar
in spirit to the RHF/BNRHF that transforms the prior ensemble in
observation space to a posterior ensemble using a probability integral
transform. The general idea is described in Grooms (Comp. Geo. 2022),
but the details of the implementation are different. The code uses
kernel density estimation with a biweight kernel and boundary corrections
to estimate the prior pdf. It uses Gaussian quadrature to evaluate the
prior and posterior cdf. Several likelihood options are implemented,
and the API is extensible to enable more likelihood options in the
future.