Rootfinding and Bug Fixes
This commit fixes several bugs in the initial kde_distribution_mod
module. It also implements a new inv_cdf function, set within a new
rootfinding_mod module. I did this because I noticed that the existing
inv_cdf function does not handle upper and lower bounds accurately.
Specifically, when finding x such that cdf(x) = q, it will return
values of x that are outside the bounds.

The code in this commit does not produce accurate cdf inversions
near the boundaries q=0,1 when there are no boundaries, or when
the boundaries are far from the data. I believe that this is
because the biweight kernel goes rapidly to zero at the edge
of its support.