Epanechnikov update
This commit switches from the biweight kernel to the Epanechnikov
kernel. The old code has been moved to kde_biweight_mod. The
change was motivated by the fact that the cdf inversion is less
accurate near the edge of the support of the pdf in unbounded
cases. In these cases the slope of the cdf (= the pdf) goes to
zero at the edge, and rootfinding does not perform well when the
slope of the function is very shallow (or zero) near the root.
The Epanechnikov kernel goes to zero less rapidly than the biweight
kernel, which makes the cdf inversion more robust near the edges.