Incorporating KDE into DART
This commit does several things to incorporate the kde distribution
module into DART. An option to use the KDE filter is added within
assim_tools, though it is forced to use the truncated-normal obs
distribution. This leads to some changes in kde_distribution_mod
to incorporate and test the truncated-normal obs distribution. The
option to use the kde cdf in the PPI transform is also added to
probit_transform_mod.